#359
- file-icon: учтены исправления косяков в дизайн-гайдах (minor)
- dropdown-icon: мелкие правки раздела (patch)
