+++ src/dropdown-item/index.ts
@@ -1,0 +1,2 @@
+export type { DropdownItemProps, DropdownItemSize } from './types';
+export { DropdownItem } from './dropdown-item';
