+++ src/file-icon/index.ts
@@ -1,0 +1,2 @@
+export type { FileIconProps, FileIconStyle } from './types';
+export { FileIcon } from './file-icon';
